patterns/rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "langchain[openai]>=1.2.10",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
]
